Finished CRUD operations for Projects route
Finished 2 gets, post, put, delete for /api/projects. Ran into unresolved blocker where I am unable to do a conditional for GET on /api/projects/:id because the falsey for this will result in a function running on an undefined entity, instead of properly returning null. I would have to change the helpers to fix this. I will come back to this if I have time.